feat: add apy tooltip
